Added all animation controls to bottom bar and removed.
